// docs/src/generated/component-api.ts
export const generatedComponentApi = {
// ...
  "aside-menu": {
    "props": [
      {
        "name": "defaultActive",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "defaultOpeneds",
        "type": "string[] | undefined",
        "required": false
      },
      {
        "name": "uniqueOpened",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "router",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "menuTrigger",
        "type": "EpPropMergeType<StringConstructor, \"click\" | \"hover\", unknown> | undefined",
        "required": false
      },
      {
        "name": "backgroundColor",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "textColor",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "activeTextColor",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "closeOnClickOutside",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "collapseTransition",
        "type": "EpPropMergeType<BooleanConstructor, unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "ellipsis",
        "type": "EpPropMergeType<BooleanConstructor, unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "popperOffset",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "ellipsisIcon",
        "type": "EpPropMergeType<(new (...args: any[]) => (string | Component) & {}) | (() => string | Component) | ((new (...args: any[]) => (string | Component) & {}) | (() => string | Component) | null)[], unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "popperEffect",
        "type": "EpPropMergeType<(new (...args: any[]) => string) | (() => PopperEffect) | ((new (...args: any[]) => string) | (() => PopperEffect) | null)[], unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "popperClass",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "popperStyle",
        "type": "EpPropMergeType<(new (...args: any[]) => string | CSSProperties) | (() => string | CSSProperties) | ((new (...args: any[]) => string | CSSProperties) | (() => string | CSSProperties) | null)[], unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "showTimeout",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "hideTimeout",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "persistent",
        "type": "EpPropMergeType<BooleanConstructor, unknown, unknown> | undefined",
        "required": false
      },
      {
        "name": "collapse",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "width",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "collapseWidth",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "theme",
        "type": "GaAsideMenuTheme | undefined",
        "required": false
      }
    ],
    "events": [
      {
        "name": "update:collapse",
        "type": "event",
        "required": false,
        "parameters": "collapse: boolean"
      },
      {
        "name": "toggle",
        "type": "event",
        "required": false,
        "parameters": "collapse: boolean"
      },
      {
        "name": "select",
        "type": "event",
        "required": false,
        "parameters": "index: string, indexPath: string[], item: MenuItemClicked, routerResult?: Promise<unknown>"
      },
      {
        "name": "open",
        "type": "event",
        "required": false,
        "parameters": "index: string, indexPath: string[]"
      },
      {
        "name": "close",
        "type": "event",
        "required": false,
        "parameters": "index: string, indexPath: string[]"
      }
    ],
    "slots": [
      {
        "name": "header",
        "type": "GaAsideMenuSlotProps",
        "required": false
      },
      {
        "name": "default",
        "type": "GaAsideMenuSlotProps",
        "required": false
      },
      {
        "name": "footer",
        "type": "GaAsideMenuSlotProps",
        "required": false
      },
      {
        "name": "collapse",
        "type": "GaAsideMenuToggleSlotProps",
        "required": false
      }
    ],
    "expose": [
      {
        "name": "menuRef",
        "type": "MenuInstance | undefined",
        "required": true
      },
      {
        "name": "toggle",
        "type": "() => void",
        "required": true
      }
    ]
  },
// ...
  "table-pagination": {
    "props": [
      {
        "name": "data",
        "type": "GaTableRow[] | undefined",
        "required": false
      },
      {
        "name": "size",
        "type": "\"\" | \"default\" | \"small\" | \"large\" | undefined",
        "required": false
      },
      {
        "name": "fit",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "stripe",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "border",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "rowKey",
        "type": "GaTableRowKey<GaTableRow> | undefined",
        "required": false
      },
      {
        "name": "showHeader",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "highlightCurrentRow",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "emptyText",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "columns",
        "type": "GaTableColumn<GaTableRow>[] | undefined",
        "required": false
      },
      {
        "name": "loading",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "loadingText",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "layout",
        "type": "string | undefined",
        "required": false
      },
      {
        "name": "currentPage",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "pageSize",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "total",
        "type": "number | undefined",
        "required": false
      },
      {
        "name": "pageSizes",
        "type": "number[] | undefined",
        "required": false
      },
      {
        "name": "background",
        "type": "boolean | undefined",
        "required": false
      },
      {
        "name": "position",
        "type": "\"center\" | \"left\" | \"right\" | undefined",
        "required": false
      },
      {
        "name": "tableTheme",
        "type": "GaTableTheme | undefined",
        "required": false
      },
      {
        "name": "paginationTheme",
        "type": "GaPaginationTheme | undefined",
        "required": false
      }
    ],
    "events": [],
    "slots": [
      {
        "name": "default",
        "type": "无作用域参数",
        "required": false
      },
      {
        "name": "empty",
        "type": "无作用域参数",
        "required": false
      },
      {
        "name": "append",
        "type": "无作用域参数",
        "required": false
      }
    ],
    "expose": []
  }
} as const